got linked sections done
The linked sections for some courses are not well accounted for. This proved to be a significantly less headache than realizing .forEach, .every, and for (;;) have custom logic for a return statement and DO NOT RETURN TO THE FUNCTION END.

I am still debating the best way for the user to manage linked sections. It's a bit of a mess right now.